Added HeatMap View
Added new heatmap view using Skia
