+++ MauiSampleApp/MainPage.xaml
@@ -12,6 +12,7 @@
         <Button Margin="30,5,30,5" Text="Color Picker" Clicked="OnColorPickerClicked" />
         <Button Margin="30,5,30,5" Text="Wizard Control" Clicked="OnWizardPageClicked" />
         <Button Margin="30,5,30,5" Text="Segmented Control" Clicked="OnSegmentedControlPageClicked" />
+        <Button Margin="30,5,30,5" Text="Heatmap" Clicked="OnHeatmapPageClicked" />
     </VerticalStackLayout>
     
 </ContentPage>
