+++ MauiSampleApp/MainPage.xaml.cs
@@ -62,6 +62,14 @@
             await Navigation.PushModalAsync(new NavigationPage(dlg));
 
         }
+        
+        private async void OnHeatmapPageClicked(object sender, EventArgs e)
+        {
+            var dlg = new HeatMapPage();
+
+            await Navigation.PushModalAsync(new NavigationPage(dlg));
+
+        }
     }
 
 }
